# Extract AbstractRaptorTransfer to improve performance