--- a/src/main/java/org/opentripplanner/routing/algorithm/raptoradapter/transit/AccessEgress.java
+++ b/src/main/java/org/opentripplanner/routing/algorithm/raptoradapter/transit/AccessEgress.java
@@ -2,19 +2,10 @@ package org.opentripplanner.routing.algorithm.raptoradapter.transit;
 
 import org.opentripplanner.routing.algorithm.raptoradapter.transit.cost.RaptorCostConverter;
 import org.opentripplanner.routing.core.State;
-import org.opentripplanner.transit.raptor.api.transit.RaptorTransfer;
+import org.opentripplanner.transit.raptor.api.transit.AbstractRaptorTransfer;
 import org.opentripplanner.util.lang.ToStringBuilder;
 
-public class AccessEgress implements RaptorTransfer {
-
-  /**
-   * "To stop" in the case of access, "from stop" in the case of egress.
-   */
-  private final int toFromStop;
-
-  private final int durationInSeconds;
-
-  private final int generalizedCost;
+public class AccessEgress extends AbstractRaptorTransfer {
 
   /**
    * This should be the last state both in the case of access and egress.
@@ -22,25 +13,12 @@ public class AccessEgress implements RaptorTransfer {
   private final State lastState;
 
   public AccessEgress(int toFromStop, State lastState) {
-    this.toFromStop = toFromStop;
-    this.durationInSeconds = (int) lastState.getElapsedTimeSeconds();
-    this.generalizedCost = RaptorCostConverter.toRaptorCost(lastState.getWeight());
+    super(
+      toFromStop,
+      (int) lastState.getElapsedTimeSeconds(),
+      RaptorCostConverter.toRaptorCost(lastState.getWeight())
+    );
     this.lastState = lastState;
-  }
-
-  @Override
-  public int stop() {
-    return toFromStop;
-  }
-
-  @Override
-  public int generalizedCost() {
-    return generalizedCost;
-  }
-
-  @Override
-  public int durationInSeconds() {
-    return durationInSeconds;
   }
 
   @Override