--- a/src/main/java/org/opentripplanner/routing/algorithm/raptoradapter/transit/request/TransferWithDuration.java
+++ b/src/main/java/org/opentripplanner/routing/algorithm/raptoradapter/transit/request/TransferWithDuration.java
@@ -1,38 +1,19 @@
 package org.opentripplanner.routing.algorithm.raptoradapter.transit.request;
 
 import org.opentripplanner.routing.algorithm.raptoradapter.transit.Transfer;
-import org.opentripplanner.transit.raptor.api.transit.RaptorTransfer;
+import org.opentripplanner.transit.raptor.api.transit.AbstractRaptorTransfer;
 
-public class TransferWithDuration implements RaptorTransfer {
-
-  private final int durationSeconds;
-  private final int cost;
+public class TransferWithDuration extends AbstractRaptorTransfer {
 
   private final Transfer transfer;
 
   public TransferWithDuration(Transfer transfer, int durationSeconds, int cost) {
+    super(transfer.getToStop(), durationSeconds, cost);
     this.transfer = transfer;
-    this.durationSeconds = durationSeconds;
-    this.cost = cost;
   }
 
   public Transfer transfer() {
     return transfer;
-  }
-
-  @Override
-  public int stop() {
-    return transfer.getToStop();
-  }
-
-  @Override
-  public int generalizedCost() {
-    return cost;
-  }
-
-  @Override
-  public int durationInSeconds() {
-    return this.durationSeconds;
   }
 
   @Override