### Summary

While profiling, I noticed that we spend quite a lot of time querying three fields from `RaptorTransfer`s. Currently we have multiple implementations of this interface, which makes `itable stub` quite a prominent part of the profile. This PR extracts the three fields used into an abstract class, which defines the getters as final, making it possible for JDK to create optimized bytecode.

```
Before

Round 1

Worker: 
 ==> mc_destination : [  490,  478,  487,  473,  483,  472,  467,  469,  474,  478 ] Avg: 477.1  (σ=7.2)

Total:  
 ==> mc_destination : [  738,  709,  731,  694,  711,  711,  697,  683,  688,  700 ] Avg: 706.2  (σ=16.8)

Round 2

Worker: 
 ==> mc_destination : [  472,  471,  471,  472,  470,  471,  469,  470,  470,  470 ] Avg: 470.6  (σ=0.9)

Total:  
 ==> mc_destination : [  704,  687,  687,  688,  686,  686,  684,  685,  685,  685 ] Avg: 687.7  (σ=5.6)



After

Round 1

Worker: 
 ==> mc_destination : [  424,  406,  405,  404,  400,  405,  409,  403,  400,  405 ] Avg: 406.1  (σ=6.5)

Total:  
 ==> mc_destination : [  669,  635,  626,  630,  631,  627,  632,  628,  631,  627 ] Avg: 633.6  (σ=12.1)

Round 2

Worker: 
 ==> mc_destination : [  414,  414,  410,  412,  410,  409,  408,  412,  410,  413 ] Avg: 411.2  (σ=2.0)

Total:  
 ==> mc_destination : [  647,  631,  629,  631,  628,  628,  627,  628,  626,  629 ] Avg: 630.4  (σ=5.7)


```